requirements.txt
flask
line-bot-sdk
requests
gunicorn
google-generativeai